move notification actions to appActions folder
